Move tests related to last-saved to test/client/feature/last-saved.spec.js
